Divide uavcan_node.c into smaller files and refactor UavcanNode's public interface
