+++ CMakeLists_app.txt
@@ -25,11 +25,35 @@ else()
   set(APP_DIR app)
 endif()
 
-set(APP_INCLUDE ${APP_DIR})
+set(UAVCAN_NODE_DIR ${APP_DIR}/uavcan_node)
+
+set(APP_INCLUDE ${APP_DIR} ${UAVCAN_NODE_DIR})
 
 set(APP_SOURCES
   ${APP_DIR}/app.h
   ${APP_DIR}/app.c
-  ${APP_DIR}/uavcan_node.h
-  ${APP_DIR}/uavcan_node.c
 )
+
+list(APPEND APP_SOURCES
+  ${UAVCAN_NODE_DIR}/uavcan_node.h
+  ${UAVCAN_NODE_DIR}/uavcan_node.c
+  ${UAVCAN_NODE_DIR}/uavcan_node_ao.h
+  ${UAVCAN_NODE_DIR}/uavcan_node_ao.c
+  ${UAVCAN_NODE_DIR}/uavcan_node_callbacks.h
+  ${UAVCAN_NODE_DIR}/uavcan_node_callbacks.c
+  ${UAVCAN_NODE_DIR}/uavcan_node_handlers.h
+  ${UAVCAN_NODE_DIR}/uavcan_node_handlers.c
+  ${UAVCAN_NODE_DIR}/uavcan_node_helpers.h
+  ${UAVCAN_NODE_DIR}/uavcan_node_helpers.c
+)
+
+# The following files are "included" in "uavcan_node.c".
+#
+# This is necessary, as we wish to have smaller .c files while maintaining
+# the benefits of single module compiler optimizations and automatic static
+# function inlining.
+#
+set_source_files_properties(${UAVCAN_NODE_DIR}/uavcan_node_ao.c PROPERTIES HEADER_FILE_ONLY TRUE)
+set_source_files_properties(${UAVCAN_NODE_DIR}/uavcan_node_callbacks.c PROPERTIES HEADER_FILE_ONLY TRUE)
+set_source_files_properties(${UAVCAN_NODE_DIR}/uavcan_node_handlers.c PROPERTIES HEADER_FILE_ONLY TRUE)
+set_source_files_properties(${UAVCAN_NODE_DIR}/uavcan_node_helpers.c PROPERTIES HEADER_FILE_ONLY TRUE)
